loginApp/app/requirements.txt
antiorm==1.2.1
click==8.1.2
colorama==0.4.4
db==0.1.1
db-sqlite3==0.0.1
Flask==2.1.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
PyJWT==2.3.0
Werkzeug==2.1.1
zipp==3.8.0
